[Patchset 13] Remove warning suppressions and fix many casting bugs
This change removes most warning suppressions, in particular
from production code. In the past we inherit a bunch of -Wno-xxx
suppressions required to build gtest and libprotobuf headers.
Doing so, however, caused the suppressions to propagate back to
the translation units that were including any protobuf header
or any auto-generated .pb.h stub.
This change moves the gtest and probobuf header to be a
system include (-isystem vs -I). Doing so implicitly blacklists
any compiler warning on the headers.
This CL then re-enables warnings and deals with the fall out of
fixes that came out of this.

Bug: 77316877
Test: pefetto_unittests / perfetto_integrationtests

+++ src/tracing/core/shared_memory_arbiter_impl.h
@@ -116,7 +116,7 @@ class SharedMemoryArbiterImpl : public SharedMemoryArbiter {
   // --- Begin lock-protected members ---
   std::mutex lock_;
   SharedMemoryABI shmem_abi_;
-  uint32_t page_idx_ = 0;
+  size_t page_idx_ = 0;
   std::unique_ptr<CommitDataRequest> commit_data_req_;
   size_t bytes_pending_commit_ = 0;  // SUM(chunk.size() : commit_data_req_).
   IdAllocator<WriterID> active_writer_ids_;
